abstract out exception handling
